Lifetime

+++ alobe/periodicEvent.cpp
@@ -3,7 +3,7 @@
 /**
  * Constructor
  */
-PeriodicEvent::PeriodicEvent(unsigned int step, unsigned int period):
+PeriodicEvent::PeriodicEvent(unsigned long step, unsigned long period):
     Event(step),
     my_triggerPeriod(period)
 {
@@ -12,7 +12,7 @@ PeriodicEvent::PeriodicEvent(unsigned int step, unsigned int period):
 /**
  * Change event filter rules (add periodicity)
  */
-void PeriodicEvent::action(Actor & actor, unsigned int step)
+void PeriodicEvent::action(Actor & actor, unsigned long step)
 {
     if (step >= my_triggerStep & (step - my_triggerStep) % my_triggerPeriod == 0) {
         this->filteredAction(actor);
